+++ src/main/java/com/example/services/CargoService.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public interface CargoService {
 
-    CargoDto findById(int cargoId);
+    CargoDto findById(Long cargoId);
     List<CargoDto> findAll();
     CargoDto getCargoByDriverId(int driverId);
     CargoDto getCargoByTruckId(int truckId);
     boolean addCargo(CargoDto cargoDto);
     boolean updateCargo(CargoDto cargoDto);
-    boolean setAcceptStatus(int cargoId, int driverId);
-    boolean setRefuseStatus(int cargoId, int driverId);
-    boolean setDeliverStatus(int cargoId, int driverId);
+    boolean setAcceptStatus(Long cargoId, int driverId);
+    boolean setRefuseStatus(Long cargoId, int driverId);
+    boolean setDeliverStatus(Long cargoId, int driverId);
 }
